infrastructure/lims_rx_endpoint/project/requirements.txt
selenium
pytest
flake8

requests
jsonpath


Django
djangorestframework
django-crispy-forms
django-tables2
django-pandas

httpie
